verified singleton container works
